requirements.txt
streamlit
Pillow
python-dotenv
requests
google-generativeai
